Generalise and cleanup list toolbar
